requirements.txt
langchain
chromadb
sentence-transformers
fastapi
uvicorn
python-docx
pdfplumber
python-pptx
pandas
